Flask/requirements.txt
jsonschema==3.2.0
matplotlib==3.3.3
numpy==1.18.5
pafy==0.5.5
Pillow==8.0.1
scipy==1.4.1
youtube-dl==2020.12.14
Flask==1.1.2
Flask-Cors==3.0.9
gunicorn==20.0.4